requirements.txt
python-dotenv==1.0.0
openai==1.3.0
PyPDF2==3.0.1
langchain==0.1.0
chromadb==0.4.18
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6